+++ src/VaultProspector.App/ProductIdentity.cs
@@ -1,0 +1,6 @@
+namespace VaultProspector.App;
+
+public static class ProductIdentity
+{
+    public const string DefaultClientId = "221af888-1c16-4637-9d45-b6dd2e1e7634";
+}